requirements.txt
# Core ML dependencies
mlx>=0.3.0
#cartesia-mlx>=0.1.0  # Replace with actual version

# Server dependencies
flask>=2.0.0
jinja2>=3.0.0

# Hugging Face dependencies
huggingface-hub>=0.19.0

# Utility dependencies
pandas>=1.3.0
numpy>=1.20.0
psutil>=5.9.0
pyopenssl>=23.0.0
python-dotenv>=1.0.0

# Visualization (if needed)
matplotlib>=3.5.0
seaborn>=0.11.0
